+++ src/CatenaLogic.Analyzers.Tests/CL0003/CL0003DiagnosticFacts.cs
@@ -4,21 +4,22 @@
     using NUnit.Framework;
 
     [TestFixture]
-    public class CL0002AnalyzerFacts
+    public class CL0003DiagnosticFacts
     {
-        private static readonly ExpectedDiagnostic ExpectedDiagnostic = ExpectedDiagnostic.Create(Descriptors.CL0002_UseAsyncSuffixForAsyncMethods);
+        private static readonly ExpectedDiagnostic ExpectedDiagnostic = ExpectedDiagnostic.Create(Descriptors.CL0003_DontUseExtensionsNamespace);
 
         [Test]
         public void Invalid_Code_01()
         {
             var before = @"
-using System;
+namespace CatenaLogic.Analyzers.↓Extensions
+{
 using System.IO;
 using System.Threading.Tasks;
 
 public class C
 {
-    public async Task ↓MyMethod()
+    public async Task MyMethod()
     {
         using (var fileStream = File.OpenRead(""filename""))
         {
@@ -26,21 +27,77 @@ public class C
                 var text = await reader.ReadToEndAsync();
             }
         }
-    }";
-            Solution.Verify<MethodsAnalyzer>(analyzer => RoslynAssert.Diagnostics(analyzer, ExpectedDiagnostic, before));
+    }
+}";
+
+            Solution.Verify<NamespacesAnalyzer>(analyzer => RoslynAssert.Diagnostics(analyzer, ExpectedDiagnostic, before));
         }
 
         [Test]
+        // Long namespaces
+        public void Invalid_Code_02()
+        {
+            var before = @"
+namespace CatenaLogic.Analyzers.Analyzers2.Analyzers3.Analyzers4.↓Extensions
+{
+using System.IO;
+using System.Threading.Tasks;
+
+public class C
+{
+    public async Task MyMethod()
+    {
+        using (var fileStream = File.OpenRead(""filename""))
+        {
+                var reader = new StreamReader(fileStream);
+                var text = await reader.ReadToEndAsync();
+            }
+        }
+    }
+}";
+
+            Solution.Verify<NamespacesAnalyzer>(analyzer => RoslynAssert.Diagnostics(analyzer, ExpectedDiagnostic, before));
+        }
+
+        [Test]
+        // Not-nested name
+        public void Invalid_Code_03()
+        {
+            var before = @"
+namespace ↓Extension
+{
+using System.IO;
+using System.Threading.Tasks;
+
+public class C
+{
+    public async Task MyMethod()
+    {
+        using (var fileStream = File.OpenRead(""filename""))
+        {
+                var reader = new StreamReader(fileStream);
+                var text = await reader.ReadToEndAsync();
+            }
+        }
+    }
+}";
+
+            Solution.Verify<NamespacesAnalyzer>(analyzer => RoslynAssert.Diagnostics(analyzer, ExpectedDiagnostic, before));
+        }
+
+        [Test]
+        // Not-nested name
         public void Valid_Code_01()
         {
             var before = @"
-using System;
+namespace ExtendedAnalyzers
+{
 using System.IO;
 using System.Threading.Tasks;
 
 public class C
 {
-    public async Task MyMethodAsync()
+    public async Task MyMethod()
     {
         using (var fileStream = File.OpenRead(""filename""))
         {
@@ -48,57 +105,10 @@ public class C
                 var text = await reader.ReadToEndAsync();
             }
         }
-    }";
-            Solution.Verify<MethodsAnalyzer>(analyzer => RoslynAssert.Valid(analyzer, before));
-        }
+    }
+}";
 
-        [Test]
-        public void Valid_Code_02()
-        {
-            var before = @"
-using System;
-using System.IO;
-using System.Threading.Tasks;
-
-public class C
-{
-    public async void MyMethod()
-    {
-        using (var fileStream = File.OpenRead(""filename""))
-        {
-                var reader = new StreamReader(fileStream);
-                var text = await reader.ReadToEndAsync();
-            }
-        }
-    }";
-
-            Solution.Verify<MethodsAnalyzer>(analyzer => RoslynAssert.Valid(analyzer, before));
-        }
-
-        [Test]
-        // Ingore Main check
-        public void Valid_Code_03()
-        {
-            var before = @"
-using System;
-using System.IO;
-using System.Threading.Tasks;
-
-class Program
-    {
-        static async Task Main(string[] args)
-        {
-            Console.WriteLine(""Hello World!"");
-
-            using (var fileStream = File.OpenRead(""filename""))
-            {
-                var reader = new StreamReader(fileStream);
-                var text = await reader.ReadToEndAsync();
-            }
-        }
-    }";
-
-            Solution.Verify<MethodsAnalyzer>(analyzer => RoslynAssert.Valid(analyzer, before));
+            Solution.Verify<NamespacesAnalyzer>(analyzer => RoslynAssert.Valid(analyzer, before));
         }
     }
 }
