+++ src/CatenaLogic.Analyzers/Analyzers/Diagnostics/CL0003/CL0003Diagnostic.cs
@@ -6,7 +6,7 @@
     using Microsoft.CodeAnalysis.CSharp.Syntax;
     using Microsoft.CodeAnalysis.Diagnostics;
 
-    public class CL0003Analyzer : ProhibitedNamespaceDiagnosticRule
+    public class CL0003Diagnostic : ProhibitedNamespaceDiagnostic
     {
         private readonly static string HandledNamespace = "Extension";
         private readonly static string HandledNamespacePlural = "Extensions";
